+++ src/RentIT.ReviewsService/ReviewService.Core/RabbitMQ/Messages/ReviewCreated.cs
@@ -1,3 +1,3 @@
 ﻿namespace ReviewService.Core.RabbitMQ.Messages;
 
-public record ReviewCreated(Guid EquipmentId, double Rating);
+public record ReviewCreated(Guid EquipmentId, decimal Rating);
