+++ src/RentIT.ReviewsService/ReviewService.Core/RabbitMQ/Messages/ReviewUpdated.cs
@@ -1,3 +1,3 @@
 ﻿namespace ReviewService.Core.RabbitMQ.Messages;
 
-public record ReviewUpdated(Guid EquipmentId, double NewRating, double OldRating);
+public record ReviewUpdated(Guid EquipmentId, decimal NewRating, decimal OldRating);
